pref: 优化日志管理模块
