+++ .idea/workspace.xml
@@ -5,7 +5,25 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="a6eb86d6-daaf-400b-b6fb-d0f01e68712a" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/web/logServer.go" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/web/model.go" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/bin/go_build_ScanTodo.exe" beforeDir="false" afterPath="$PROJECT_DIR$/bin/go_build_ScanTodo.exe" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.go" beforeDir="false" afterPath="$PROJECT_DIR$/main.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/scan/ping/ping.go" beforeDir="false" afterPath="$PROJECT_DIR$/scan/ping/ping.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/scan/scan.go" beforeDir="false" afterPath="$PROJECT_DIR$/scan/scan.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/log_h.go" beforeDir="false" afterPath="$PROJECT_DIR$/scanLog/log_h.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/Ping日志_2023_04_16.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/Ping日志_2023_04_22.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/Ping日志_2023_04_23.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/Ping日志_2023_04_24.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/Ping日志_2023_04_25.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/全局日志_2023_04_11.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/全局日志_2023_04_12.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/全局日志_2023_04_23.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/扫描日志_2023_04_11.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/scanLog/logs/扫描日志_2023_04_12.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/webServer.go" beforeDir="false" afterPath="$PROJECT_DIR$/web/webServer.go" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -53,7 +71,13 @@
     "nodejs_package_manager_path": "npm"
   }
 }]]></component>
-  <component name="RunManager" selected="Go Build.go build ScanTodo/scan/ping/cmd">
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="D:\code\ScanTodo\web" />
+      <recent name="D:\code\ScanTodo\release" />
+    </key>
+  </component>
+  <component name="RunManager" selected="Go Build.go build ScanTodo">
     <configuration default="true" type="GoApplicationRunConfiguration" factoryName="Go Application">
       <module name="ScanTodo" />
       <working_directory value="$PROJECT_DIR$" />
