+++ scanLog/logs/HTTPLogPath/2023_04_26.txt
@@ -1,0 +1,5 @@
+[debug]2023/04/26 16:51:22 main.go:32: 全局日志开始启动.............................
+[debug]2023/04/26 16:51:22 main.go:37: 服务启动中,全局日志加载完成
+[debug]2023/04/26 16:51:22 main.go:41: 服务启动中,websocket实例初始化完成
+[debug]2023/04/26 16:51:22 main.go:43: 服务启动中,websocket开启监听完成
+[debug]2023/04/26 16:51:22 main.go:48: 服务启动成功,功能点加载完成
