--- scanLog/logs/Ping日志_2023_04_16.txt
@@ -1,1 +1,0 @@
-[error]2023/04/16 13:05:27 ping.go:104: 目标ip为空
