--- scanLog/logs/Ping日志_2023_04_22.txt
@@ -1,4 +1,0 @@
-[info]2023/04/22 15:57:30 cmdTodo.go:71: 开始ping %s (%s): 
-  192.168.0.0 192.168.0.0
-[info]2023/04/22 15:58:08 cmdTodo.go:71: 开始ping 192.168.0.0 (192.168.0.0): 
-
