--- scanLog/logs/全局日志_2023_04_11.txt
@@ -1,27 +1,0 @@
-[debug]2023/04/11 22:09:23 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:09:23 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:09:23 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:12:40 main.go:27: 全局日志启动中.............................
-[debug]2023/04/11 22:12:40 main.go:35: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:12:40 main.go:39: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:12:40 main.go:41: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:14:26 main.go:27: 全局日志开始启动.............................
-[debug]2023/04/11 22:14:26 main.go:35: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:14:26 main.go:39: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:14:26 main.go:41: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:14:26 main.go:47: 服务启动完成,功能点加载完成
-[debug]2023/04/11 22:15:13 main.go:27: 全局日志开始启动.............................
-[debug]2023/04/11 22:15:13 main.go:35: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:15:13 main.go:39: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:15:13 main.go:41: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:15:13 main.go:47: 服务启动完成,功能点加载完成
-[debug]2023/04/11 22:15:31 main.go:27: 全局日志开始启动.............................
-[debug]2023/04/11 22:15:31 main.go:35: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:15:31 main.go:39: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:15:31 main.go:41: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:15:31 main.go:47: 服务启动完成,功能点加载完成
-[debug]2023/04/11 22:15:49 main.go:27: 全局日志开始启动.............................
-[debug]2023/04/11 22:15:49 main.go:35: 服务启动中,全局日志加载完成
-[debug]2023/04/11 22:15:49 main.go:39: 服务启动中,websocket实例初始化完成
-[debug]2023/04/11 22:15:49 main.go:41: 服务启动中,websocket开启监听完成
-[debug]2023/04/11 22:15:49 main.go:47: 服务启动完成,功能点加载完成
