--- scanLog/logs/全局日志_2023_04_12.txt
@@ -1,50 +1,0 @@
-[debug]2023/04/12 20:16:03 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:16:03 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:16:03 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:16:03 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:16:03 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:31:23 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:31:23 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:31:23 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:31:23 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:31:23 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:35:31 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:35:31 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:35:31 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:35:31 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:35:31 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:37:27 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:37:27 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:37:27 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:37:27 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:37:27 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:40:27 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:40:27 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:40:27 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:40:27 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:40:27 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:42:59 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:42:59 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:42:59 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:42:59 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:42:59 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:46:09 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:46:09 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:46:09 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:46:09 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:46:09 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:47:06 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:47:06 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:47:06 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:47:06 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:47:06 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:54:19 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:54:19 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:54:19 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:54:19 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:54:19 main.go:46: 服务启动成功,功能点加载完成
-[debug]2023/04/12 20:56:12 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/12 20:56:12 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/12 20:56:12 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/12 20:56:12 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/12 20:56:12 main.go:46: 服务启动成功,功能点加载完成
