--- scanLog/logs/全局日志_2023_04_23.txt
@@ -1,5 +1,0 @@
-[debug]2023/04/23 17:23:19 main.go:26: 全局日志开始启动.............................
-[debug]2023/04/23 17:23:19 main.go:34: 服务启动中,全局日志加载完成
-[debug]2023/04/23 17:23:19 main.go:38: 服务启动中,websocket实例初始化完成
-[debug]2023/04/23 17:23:19 main.go:40: 服务启动中,websocket开启监听完成
-[debug]2023/04/23 17:23:19 main.go:46: 服务启动成功,功能点加载完成
