--- scanLog/logs/扫描日志_2023_04_11.txt
@@ -1,7 +1,0 @@
-[warn]2023/04/11 22:11:19 TcpScan.go:42: ------------------启动----------------------
-[info]2023/04/11 22:11:19 TcpScan.go:44: 准备扫描的ip数量: 1
-[info]2023/04/11 22:11:19 TcpScan.go:65: ip go程 1 个
-[info]2023/04/11 22:11:19 TcpScan.go:66: 每个ip准备访问的端口 go程 1 个
-[info]2023/04/11 22:11:19 TcpScan.go:67: 当前超时时间设置为 1000 毫秒
-[info]2023/04/11 22:11:19 TcpScan.go:91: tcp扫描结束
-[warn]2023/04/11 22:11:19 TcpScan.go:47: -------------------结束---------------------
