--- scanLog/logs/扫描日志_2023_04_12.txt
@@ -1,21 +1,0 @@
-[error]2023/04/12 20:33:03 scan.go:58: 无法实现 Icmp 这个类型!!!!!!!!!
-[error]2023/04/12 20:35:33 webServer.go:82: icmp start error 源ip:  格式错误
-[error]2023/04/12 20:36:29 webServer.go:82: icmp start error 源ip: 127.0.0.1-127.0.0.20 格式错误
-[error]2023/04/12 20:36:36 webServer.go:82: icmp start error 目标ip: 127.0.0.1 格式错误
-[error]2023/04/12 20:36:52 webServer.go:82: icmp start error 目标ip: 127.0.0.1 格式错误
-[error]2023/04/12 20:38:28 IcmpScan.go:52: 打开ip4:icmp连接错误 dial ip4:icmp 127.0.0.1->192.168.232.141: connect: A socket operation was attempted to an unreachable network.
-[error]2023/04/12 20:38:28 webServer.go:82: icmp start error dial ip4:icmp 127.0.0.1->192.168.232.141: connect: A socket operation was attempted to an unreachable network.
-[info]2023/04/12 20:39:16 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:40:29 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:41:37 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:42:08 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:42:27 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:43:09 IcmpScan.go:77: send icmp packet success
-[info]2023/04/12 20:44:19 IcmpScan.go:77: send icmp packet success
-[info]2023/04/12 20:46:12 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:47:10 IcmpScan.go:76: send icmp packet success
-[error]2023/04/12 20:54:32 IcmpScan.go:53: 打开ip4:icmp连接错误 dial ip4:icmp 192.168.0.14->192.168.232.141: bind: The requested address is not valid in its context.
-[error]2023/04/12 20:54:32 webServer.go:82: icmp start error dial ip4:icmp 192.168.0.14->192.168.232.141: bind: The requested address is not valid in its context.
-[info]2023/04/12 20:54:40 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:55:50 IcmpScan.go:76: send icmp packet success
-[info]2023/04/12 20:56:16 IcmpScan.go:75: send icmp packet success
